Cleaned up the drake migration script a bit.
