+++ lib/models/apis/github.js
@@ -807,48 +807,3 @@ Github.prototype.addDeployKeyIfNotAlready = function (repo, cb) {
     })
   })
 }
-
-/**
- * Updates old github webhook urls with the new drake url.
- * @param  {string}   shortRepo Repo with hooks to update
- * @param  {Function} cb        Called when the hook has been updated.
- */
-Github.prototype.updateOldRepoHookURL = function (shortRepo, cb) {
-  var log = logger.log.child({ tx: true, shortRepo: shortRepo })
-  var self = this
-  var oldHookUrl = process.env.FULL_API_DOMAIN + process.env.GITHUB_HOOK_PATH
-
-  log.info('Github.prototype.updateOldRepoHookURL')
-  this._listRepoHooks(shortRepo, function (err, existingHooks) {
-    if (err) {
-      log.error({
-        err: err
-      }, 'Github.prototype.updateOldRepoHookURL error listing hooks')
-      return cb(err)
-    }
-
-    // Find a hook with the old url
-    var hook = find(existingHooks, hasKeypaths({
-      'config.url': oldHookUrl,
-      active: true,
-      'events[0]': '*'
-    }))
-
-    // If the hook doesn't exist, skip this repository.
-    if (!hook) {
-      return cb()
-    }
-
-    // Update the hook with the new github webhook url
-    self._updateRepoHook(hook.id, shortRepo, function (err) {
-      if (err) {
-        log.info({
-          err: err
-        }, 'Github.prototype.updateOldRepoHookURL hook failed to update')
-        return cb(err)
-      }
-      log.info('Github.prototype.updateOldRepoHookURL hook updated')
-      cb(null)
-    })
-  })
-}
